+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added default BCC - [#341](https://github.com/hkdb/aerion/issues/341)
 - Added default reply-to - [#404](https://github.com/hkdb/aerion/issues/404)
 - Added merge identical messages in thread - [#222](https://github.com/hkdb/aerion/issues/222)
+- Relabeled spellcheck dictionary English to English (US)
+- Added spellcheck dictionary English (UK) - [#417](https://github.com/hkdb/aerion/issues/417)
 - Fixed post delete ghost messages on message list
 - New app icon with full hicolor size set + scalable SVG - [#395](https://github.com/hkdb/aerion/issues/395) (Special thanks to @gianlucamazza) - [#99](https://github.com/hkdb/aerion/issues/99) 
 - Fixed thread sender and preview in message list - [#169](https://github.com/hkdb/aerion/issues/169)